+++ lib/src/tooling/dart_interaction.dart
@@ -14,17 +14,17 @@ abstract class DartInteraction {
   /// The decision which command to run is taken from the pubspec.yaml file in
   /// [forDirectory]
   /// This command determines which tool to use based on the pubspcec.yaml file.
-  /// If [overrideUseFlutterCommand] is given then this value will be used to
+  /// If [forceUseFlutterTool] is given then this value will be used to
   /// determine which tool to use. [true] means Flutter, [false] means Dart
   static Future runDartOrFlutterCommand(
     String forDirectory, {
     List<String> args = const [],
     StdoutSession? stdoutSession,
-    bool? overrideUseFlutterCommand,
+    bool forceUseFlutterTool = false,
   }) async {
     bool useFlutter = false;
-    if (overrideUseFlutterCommand == null) {
-      // if no override is specified we get the information which tool to use
+    if (!forceUseFlutterTool) {
+      // if we are not forced to use Flutter then we get this information
       // from the pubspec.yaml file
       final pubspecPath = path.join(forDirectory, 'pubspec.yaml');
       final pubspecExists = await File(pubspecPath).exists();
@@ -36,8 +36,8 @@ abstract class DartInteraction {
       final pubSpec = Pubspec.parse(yamlContent);
       useFlutter = pubSpec.dependencies.containsKey('flutter');
     } else {
-      // if the decision is overridden we use the given value (overrideUseFlutterCommand == true => Flutter, == false => Dart)
-      useFlutter = overrideUseFlutterCommand;
+      // here we are forced to use Flutter (forceUseFlutterTool == true)
+      useFlutter = true;
     }
     if (!useFlutter) {
       return _runDartOrFlutterCommand(