+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Version 0.19.0
-- introduces `override-use-flutter` option for all commands to override the automatic decision whether to use Flutter or Dart.
+- introduces `force-use-flutter` option for all commands to force dart_apitool to use the `flutter` command.
 
 ## Version 0.18.0
 - add missing export to json output for `extract` command